requirements.txt
fastapi[standard]
FlagEmbedding==1.3.5
google-genai==1.56.0
langchain-core==1.2.6
langchain-text-splitters==1.1.0
python-dotenv==1.2.1
qdrant-client==1.16.2
tiktoken==0.12.0
ragas
langchain-google-genai
streamlit
plotly
langchain-openai